move shared configs into rl configs
